+++ checker_test.go
@@ -9,6 +9,8 @@ import (
 
 	"net"
 
+	"time"
+
 	"github.com/sagikazarmark/healthz"
 )
 
@@ -113,7 +115,7 @@ func TestHTTPChecker_Check(t *testing.T) {
 	}))
 	defer ts.Close()
 
-	checker := healthz.NewHTTPChecker(ts.URL)
+	checker := healthz.NewHTTPChecker(ts.URL, 0)
 
 	assertCheckerSuccessful(t, checker)
 }
@@ -125,9 +127,39 @@ func TestHTTPChecker_Check_Fail(t *testing.T) {
 	}))
 	defer ts.Close()
 
-	checker := healthz.NewHTTPChecker(ts.URL)
+	checker := healthz.NewHTTPChecker(ts.URL, 0)
 
 	assertCheckerFailed(t, checker)
+}
+
+func TestHTTPChecker_Check_Timeout(t *testing.T) {
+	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		time.Sleep(5 * time.Nanosecond)
+		w.WriteHeader(http.StatusOK)
+		w.Write([]byte("ok"))
+	}))
+	defer ts.Close()
+
+	checker := healthz.NewHTTPChecker(ts.URL, 15*time.Millisecond)
+
+	assertCheckerSuccessful(t, checker)
+}
+
+func TestHTTPChecker_Check_Timeout_Fail(t *testing.T) {
+	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		time.Sleep(10 * time.Nanosecond)
+		w.WriteHeader(http.StatusOK)
+		w.Write([]byte("ok"))
+	}))
+	defer ts.Close()
+
+	checker := healthz.NewHTTPChecker(ts.URL, 3*time.Nanosecond)
+
+	err := checker.Check()
+
+	if err == nil {
+		t.Fatal("Expected error, none received")
+	}
 }
 
 func TestTCPChecker_Check(t *testing.T) {